api: Move cookie session management logic to the authenticator for more flexibility

+++ moulinette/actionsmap.py
@@ -10,6 +10,7 @@ from typing import List, Optional
 from time import time
 from collections import OrderedDict
 from importlib import import_module
+from functools import cache
 
 from moulinette import m18n, Moulinette
 from moulinette.core import (
@@ -413,6 +414,9 @@ class ActionsMap:
             # Read actions map from yaml file
             actionsmap = read_yaml(actionsmap_yml)
 
+            if not actionsmap["_global"].get("cache", True):
+                return actionsmap
+
             # Delete old cache files
             for old_cache in glob.glob(f"{actionsmap_yml_dir}/.{actionsmap_yml_file}.*.pkl"):
                 os.remove(old_cache)
@@ -456,6 +460,7 @@ class ActionsMap:
         self.extraparser = ExtraArgumentParser(top_parser.interface)
         self.parser = self._construct_parser(actionsmap, top_parser)
 
+    @cache
     def get_authenticator(self, auth_method):
 
         if auth_method == "default":
@@ -534,7 +539,7 @@ class ActionsMap:
             full_action_name = "{}.{}.{}".format(namespace, category, action)
 
         # Lock the moulinette for the namespace
-        with MoulinetteLock(namespace, timeout):
+        with MoulinetteLock(namespace, timeout, self.enable_lock):
             start = time()
             try:
                 mod = __import__(
@@ -616,7 +621,7 @@ class ActionsMap:
         _global = actionsmap.pop("_global", {})
 
         self.namespace = _global["namespace"]
-        self.cookie_name = _global["cookie_name"]
+        self.enable_lock = _global.get("lock", True)
         self.default_authentication = _global["authentication"][
             interface_type
         ]
